# lib/api-zod/src/generated/types/passActivateRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kultr Hub API
 * OpenAPI spec version: 0.2.0
 */

export interface PassActivateRequest {
  /** Reference of a payment already verified via POST /payments/pass/verify. */
  reference: string;
}